roachtest: move schemachange/tpcc test to schemachange.go
Release note: None
